refactor(module:button & card): refactor
