+++ components/card/nz-card.component.ts
@@ -1,23 +1,15 @@
-import {
-  Component,
-  ContentChild,
-  HostBinding,
-  Input,
-  TemplateRef
-} from '@angular/core';
-
-import { toBoolean } from '../core/util/convert';
-
+import { ChangeDetectionStrategy, Component, ContentChild, HostBinding, Input, TemplateRef } from '@angular/core';
+import { InputBoolean } from '../core/util/convert';
 import { NzCardTabComponent } from './nz-card-tab.component';
 
 @Component({
   selector           : 'nz-card',
   preserveWhitespaces: false,
+  changeDetection    : ChangeDetectionStrategy.OnPush,
   templateUrl        : './nz-card.component.html',
   styles             : [ `
     :host {
       display: block;
-      position: relative;
     }
   ` ],
   host               : {
@@ -26,18 +18,16 @@ import { NzCardTabComponent } from './nz-card-tab.component';
   }
 })
 export class NzCardComponent {
-  private _bordered = true;
-  private _loading = false;
-  private _hoverable = false;
-  private _title: string | TemplateRef<void>;
-  private _extra: string | TemplateRef<void>;
-  isTitleString: boolean;
-  isExtraString: boolean;
-  @ContentChild(NzCardTabComponent) tab: NzCardTabComponent;
+  @Input() @InputBoolean() @HostBinding('class.ant-card-bordered') nzBordered = true;
+  @Input() @InputBoolean() nzLoading = false;
+  @Input() @InputBoolean() @HostBinding('class.ant-card-hoverable') nzHoverable = false;
   @Input() nzBodyStyle: { [ key: string ]: string };
   @Input() nzCover: TemplateRef<void>;
   @Input() nzActions: Array<TemplateRef<void>> = [];
   @Input() nzType: string;
+  @ContentChild(NzCardTabComponent) tab: NzCardTabComponent;
+  isTitleString: boolean;
+  isExtraString: boolean;
 
   @Input()
   set nzTitle(value: string | TemplateRef<void>) {
@@ -69,32 +59,6 @@ export class NzCardComponent {
     return !!this.tab;
   }
 
-  @Input()
-  @HostBinding('class.ant-card-bordered')
-  set nzBordered(value: boolean) {
-    this._bordered = toBoolean(value);
-  }
-
-  get nzBordered(): boolean {
-    return this._bordered;
-  }
-
-  @Input()
-  set nzLoading(value: boolean) {
-    this._loading = toBoolean(value);
-  }
-
-  get nzLoading(): boolean {
-    return this._loading;
-  }
-
-  @Input()
-  @HostBinding('class.ant-card-hoverable')
-  set nzHoverable(value: boolean) {
-    this._hoverable = toBoolean(value);
-  }
-
-  get nzHoverable(): boolean {
-    return this._hoverable;
-  }
+  private _title: string | TemplateRef<void>;
+  private _extra: string | TemplateRef<void>;
 }
